Added documentation for `BestDateFinder.java`. Changed `FreeTimeFinder.java` name to `BestDateFinder.java`
